+++ src/typescript/hooks/ready.ts
@@ -1,5 +1,8 @@
 import { configureFoundryOnReady } from "../config.js";
+import { CONSTANTS } from "../constants.js";
+import { getGame } from "../foundryHelpers.js";
 import { migrateWorld, migrationNeeded } from "../migrations/world.js";
+import { initializedSettingName } from "../settings.js";
 
 /** Register system callbacks for the ready hook. */
 export default function registerForReady(): void {
@@ -10,6 +13,7 @@ export default function registerForReady(): void {
 function ready(): void {
   configureFoundryOnReady();
   migrate();
+  getGame().settings.set(CONSTANTS.systemId, initializedSettingName, true);
 }
 
 /** Run the system migrations, if needed. */