+++ src/typescript/migrations/world.ts
@@ -1,18 +1,21 @@
 import { CONSTANTS } from "../constants.js";
 import { getGame } from "../foundryHelpers.js";
-import { migrVerSettingName } from "../settings.js";
+import { initializedSettingName, migrVerSettingName } from "../settings.js";
 import migrateActors from "./actors.js";
 import migrateItems from "./items.js";
 
 /** Check if a data migration is needed and possible. */
 export function migrationNeeded(): boolean {
   if (!(game instanceof Game)) return false;
+  if (!game.settings.get(CONSTANTS.systemId, initializedSettingName))
+    return false;
   if (!game.user) return false;
   if (!game.user.isGM) return false;
 
   return isNewerVersionThanLast(CONSTANTS.needsMigrationVersion);
 }
 
+/** Check if the given version is newer than the version in the setting. */
 export function isNewerVersionThanLast(version: string): boolean {
   const lastMigrVersion = getGame().settings.get(
     CONSTANTS.systemId,
@@ -43,11 +46,8 @@ export async function migrateWorld(): Promise<void> {
   await migrateActors();
   await migrateItems();
 
-  getGame().settings.set(
-    CONSTANTS.systemId,
-    migrVerSettingName,
-    getGame().system.data.version
-  );
+  setMigrationCurrentVersion();
+
   if (ui.notifications) {
     ui.notifications.info(
       getGame().i18n.format("wv.migration.completed", {
@@ -57,4 +57,13 @@ export async function migrateWorld(): Promise<void> {
       { permanent: true }
     );
   }
+}
+
+/** Set the migration version current system version. */
+function setMigrationCurrentVersion(): void {
+  getGame().settings.set(
+    CONSTANTS.systemId,
+    migrVerSettingName,
+    getGame().system.data.version
+  );
 }